Deprecates server/client side authentication strategy. Adds support for media uploads
