+++ examples/media.js
@@ -1,6 +1,7 @@
 'use strict';
 
 var Twitter = require('../lib/twitter');
+var fs = require('fs');
 
 var client = new Twitter({
   consumer_key: process.env.TWITTER_CONSUMER_KEY,
@@ -9,4 +10,7 @@ var client = new Twitter({
   access_token_secret: process.env.TWITTER_ACCESS_TOKEN_SECRET,
 });
 
-client.post('media/upload', {}, function(err, data){});
+client.post('media/upload', {}, function(err, data){
+  console.log(err);
+  console.log(data);
+});
