+++ lib/twitter.js
@@ -8,11 +8,12 @@ var Keygrip = require('keygrip');
 var merge = require('util')._extend;
 var querystring = require('querystring');
 var urlparse = require('url').parse;
+var streamparser = require('./parser');
 
 // Package version
 var VERSION = require('../package.json').version;
 
-function Twitter(options) {
+function Twitter (options) {
   if (!(this instanceof Twitter)) return new Twitter(options);
 
   this.VERSION = VERSION;
@@ -41,6 +42,7 @@ function Twitter(options) {
     user_stream_base: 'https://userstream.twitter.com/1.1',
     site_stream_base: 'https://sitestream.twitter.com/1.1',
     filter_stream_base: 'https://stream.twitter.com/1.1/statuses',
+    media_base: 'https://upload.twitter.com/1.1',
 
     secure: false, // force use of https for login/gatekeeper
     cookie: 'twauth',
@@ -69,11 +71,22 @@ function Twitter(options) {
   );
 
   this.__generateURL = function(path, query) {
+
+    query = query === null ? {} : query;
+
     if (urlparse(path).protocol === null) {
       if (path.charAt(0) !== '/') {
         path = '/' + path;
       }
-      path = this.options.rest_base + path;
+
+      switch (true) {
+        case /^\/media/.test(path):
+          path = this.options.media_base + path;
+          break;
+        default:
+          path = this.options.rest_base + path;
+          break;
+      }
     }
 
     // Remove trailing slash
@@ -91,7 +104,6 @@ function Twitter(options) {
     }
     return path;
   };
-
 }
 
 Twitter.prototype.__request = function(method, path, params, callback) {
@@ -159,25 +171,147 @@ Twitter.prototype.__request = function(method, path, params, callback) {
 
   // Make oauth request and pass request arguments
   this.oauth[method].apply(this.oauth,request);
-
-  return this;
 };
 
 /**
  * GET
  */
 Twitter.prototype.get = function(url, params, callback) {
-  this.__request('GET', url, params, callback);
+  return this.__request('GET', url, params, callback);
 };
 
 /**
  * POST
  */
 Twitter.prototype.post = function(url, params, callback) {
-  this.__request('POST', url, params, callback);
+  return this.__request('POST', url, params, callback);
 };
 
-// Load legacy helper methods
-Twitter = require('./legacy')(Twitter);
+/**
+ * STREAM
+ */
+Twitter.prototype.stream = function (method, params, callback) {
+  if (typeof params === 'function') {
+    callback = params;
+    params = null;
+  }
+
+  var stream_base = this.options.stream_base;
+
+  // Stream type customisations
+  if (method === 'user') {
+    stream_base = this.options.user_stream_base;
+    // Workaround for node-oauth vs. twitter commas-in-params bug
+    if ( params && params.track && Array.isArray(params.track) ) {
+      params.track = params.track.join(',');
+    }
+
+  } else if (method === 'site') {
+    stream_base = this.options.site_stream_base;
+    // Workaround for node-oauth vs. twitter double-encode-commas bug
+    if ( params && params.follow && Array.isArray(params.follow) ) {
+      params.follow = params.follow.join(',');
+    }
+  } else if (method === 'filter') {
+    stream_base = this.options.filter_stream_base;
+    // Workaround for node-oauth vs. twitter commas-in-params bug
+    if ( params && params.track && Array.isArray(params.track) ) {
+      params.track = params.track.join(',');
+    }
+  }
+
+  var url = stream_base + '/' + escape(method) + '.json';
+
+  var request = this.oauth.post(url,
+    this.options.access_token_key,
+    this.options.access_token_secret,
+    params
+  );
+
+  var stream = new streamparser();
+  stream.destroy = function() {
+    // FIXME: should we emit end/close on explicit destroy?
+    if ( typeof request.abort === 'function' )
+    request.abort(); // node v0.4.0
+    else
+    request.socket.destroy();
+  };
+
+  request.on('response', function(response) {
+    // FIXME: Somehow provide chunks of the response when the stream is connected
+    // Pass HTTP response data to the parser, which raises events on the stream
+    response.on('follow', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('favorite', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('unfavorite', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('block', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('unblock', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('list_created', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('list_destroyed', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('list_updated', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('list_member_added', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('list_member_removed', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('list_user_subscribed', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('list_user_unsubscribed', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('user_update', function(chunk){
+      stream.receive(chunk);
+    });
+
+    response.on('data', function(chunk) {
+      stream.receive(chunk);
+    });
+
+    response.on('error', function(error) {
+      stream.emit('error', error);
+    });
+
+    response.on('end', function() {
+      stream.emit('end', response);
+    });
+  });
+
+  request.on('error', function(error) {
+    stream.emit('error', error);
+  });
+  request.end();
+
+  if ( typeof callback === 'function' ) callback(stream);
+  return this;
+};
 
 module.exports = Twitter;
